src/requirements.txt
datadings==3.4.4
joblib==1.1.0
matplotlib==3.5.1
numpy==1.22.2
opencv_python_headless==4.5.5.62
optuna==2.10.0
pandas==1.4.1
Pillow==9.2.0
scikit_learn==1.1.1
seaborn==0.11.2
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
